requirements.txt
pillow
torch
torchvision
numpy<2.0
matplotlib